+++ R/sugar.R
@@ -128,6 +128,40 @@ oi = function(
     callbacks = callbacks)
 }
 
+#' @title Syntactic Sugar for Optimization Test Functions
+#'
+#' @description
+#' This function complements [mlr_test_functions] with functions in the spirit
+#' of `mlr_sugar` from \CRANpkg{mlr3}.
+#'
+#' @param .key (`character(1)`)\cr
+#' Key passed to the respective [dictionary][mlr3misc::Dictionary] to retrieve
+#' the object.
+#' @param .keys (`character()`)\cr
+#' Keys passed to the respective [dictionary][mlr3misc::Dictionary] to retrieve
+#' multiple objects.
+#' @param ... (named `list()`)\cr
+#' Named arguments passed to the constructor. See
+#' [mlr3misc::dictionary_sugar_get()] for more details.
+#'
+#' @return
+#' * [ObjectiveRFun] for `otfun()`.
+#' * list of [ObjectiveRFun] for `otfuns()`.
+#'
+#' @export
+#' @examples
+#' obj = otfun("branin")
+#' obj$eval(list(x1 = 1, x2 = 2))
+otfun = function(.key, ...) {
+  dictionary_sugar_get(mlr_test_functions, .key, ...)
+}
+
+#' @rdname otfun
+#' @export
+otfuns = function(.keys, ...) {
+  dictionary_sugar_mget(mlr_test_functions, .keys, ...)
+}
+
 #' @title Syntactic Sugar for Asynchronous Optimization Instance Construction
 #'
 #' @description